requirements.txt
python-telegram-bot
flask
gunicorn
markovify